split deprecated file
