+++ deprecated.rest-struct-params.go
@@ -1,0 +1,29 @@
+package disgord
+
+
+// Deprecated use UpdateChannel
+type UpdateChannelParams = UpdateChannel
+
+// Deprecated use UpdateChannelPermissions
+type UpdateChannelPermissionsParams = UpdateChannelPermissions
+
+// Deprecated use UpdateGuildRolePositions
+type UpdateGuildRolePositionsParams = UpdateGuildRolePositions
+
+// Deprecated use UpdateGuildIntegration
+type UpdateGuildIntegrationParams = UpdateGuildIntegration
+
+// Deprecated use UpdateMember
+type UpdateMemberParams = UpdateMember
+
+// Deprecated use UpdateMessage
+type UpdateMessageParams = UpdateMessage
+
+// Deprecated use UpdateRole
+type UpdateRoleParams = UpdateRole
+
+// Deprecated use UpdateUser
+type UpdateUserParams = UpdateUser
+
+// Deprecated use UpdateWebhook
+type UpdateWebhookParams = UpdateWebhook
